+++ docs/changes.rst
@@ -9,6 +9,7 @@ Change Log
   * created :meth:`.put_d` to put the contents of a local directory to a remote one
   * created a context manager chdir method, :meth:`.cd`
   * created :meth:`.put_r` to recursively put the contents of a local directory to a remote one
+  * fixed a bug with :func:`.st_mode_to_int` on py3 (#52)
 
 * 0.2.7 (current, released 2014-05-24)